Initial commit
uhslc_format_conversion.py reads .dat files and converts/saves the data
to .nc and .csv formats. Metadata is taken from a combination of .dat
headers and the IOC SSC. ISO country names and codes were copied from
the web converted to .json for inclusion as metadata in the .nc files.
readwrite.py is a python library of classes/functions that
read/manipulate/save data.